+++ config/detekt/detekt.yml
@@ -13,6 +13,8 @@ style:
   UnusedPrivateMember:
     excludes:
       - "**/Attachment.kt"
+  ReturnCount:
+    active: false
 naming:
   MatchingDeclarationName:
     active: false
