+++ sentry-kotlin-multiplatform-gradle-plugin/src/main/java/io/sentry/kotlin/multiplatform/gradle/CocoaFrameworkLinker.kt
@@ -25,7 +25,7 @@ class CocoaFrameworkLinker(
                     "Start resolving Sentry Cocoa framework paths for target: ${target.name}"
                 )
                 processTarget(target)
-            } catch (e: Exception) {
+            } catch (e: FrameworkLinkingException) {
                 throw FrameworkLinkingException("Failed to configure ${target.name}: ${e.message}", e)
             }
         }
@@ -44,55 +44,6 @@ class CocoaFrameworkLinker(
     }
 }
 
-class FrameworkLinker(
-    private val logger: Logger
-) {
-    fun configureBinaries(
-        binaries: KotlinNativeBinaryContainer,
-        dynamicPath: String?,
-        staticPath: String?
-    ) {
-        binaries.all { binary ->
-            when (binary) {
-                is TestExecutable -> linkTestBinary(binary, chooseTestPath(dynamicPath, staticPath))
-                is Framework -> linkFrameworkBinary(binary, dynamicPath, staticPath)
-                else -> logger.info("Ignoring binary type: ${binary::class.java.simpleName}")
-            }
-        }
-    }
-
-    private fun chooseTestPath(dynamic: String?, static: String?) = when {
-        dynamic != null && static != null -> {
-            logger.debug("Both framework types available, preferring dynamic for tests")
-            dynamic
-        }
-
-        dynamic != null -> dynamic
-        static != null -> static
-        else -> throw FrameworkLinkingException("No valid framework path found for tests")
-    }
-
-    private fun linkTestBinary(binary: TestExecutable, path: String) {
-        // Linking in test binaries works with both dynamic and static framework
-        binary.linkerOpts("-rpath", path, "-F$path")
-        logger.info("Linked Sentry Cocoa framework to test binary ${binary.name}")
-    }
-
-    private fun linkFrameworkBinary(binary: Framework, dynamicPath: String?, staticPath: String?) {
-        val (path, type) = when {
-            binary.isStatic && staticPath != null -> staticPath to "static"
-            !binary.isStatic && dynamicPath != null -> dynamicPath to "dynamic"
-            else -> throw FrameworkLinkingException(
-                "Framework mismatch for ${binary.name}. " +
-                    "Required ${if (binary.isStatic) "static" else "dynamic"} Sentry Cocoa framework not found."
-            )
-        }
-
-        binary.linkerOpts("-F$path")
-        logger.info("Linked $type Sentry Cocoa framework to ${binary.name}")
-    }
-}
-
 internal class FrameworkLinkingException(
     message: String,
     cause: Throwable? = null
@@ -106,7 +57,8 @@ internal class FrameworkLinkingException(
  * across different versions. For example:
  * - iosArm64 -> [SentryCocoaFrameworkArchitectures.IOS_ARM64]
  * - macosArm64 -> [SentryCocoaFrameworkArchitectures.MACOS_ARM64_AND_X64]
- * * @return Set of possible architecture folder names for the given target. Returns empty set if target is not supported.
+ * @return Set of possible architecture folder names for the given target.
+ * Returns empty set if target is not supported.
  */
 internal fun KotlinNativeTarget.toSentryFrameworkArchitecture(): Set<String> = buildSet {
     when (name) {
@@ -130,7 +82,14 @@ internal object SentryCocoaFrameworkArchitectures {
     val WATCHOS_SIMULATOR_AND_X64 = setOf("watchos-arm64_i386_x86_64-simulator")
 
     // Used for tests
-    val all = setOf(IOS_SIMULATOR_AND_X64, IOS_ARM64, MACOS_ARM64_AND_X64, TVOS_SIMULATOR_AND_X64, TVOS_ARM64, WATCHOS_ARM, WATCHOS_SIMULATOR_AND_X64)
+    val all = setOf(
+        IOS_SIMULATOR_AND_X64,
+        IOS_ARM64, MACOS_ARM64_AND_X64,
+        TVOS_SIMULATOR_AND_X64,
+        TVOS_ARM64,
+        WATCHOS_ARM,
+        WATCHOS_SIMULATOR_AND_X64
+    )
 }
 
 internal fun KotlinMultiplatformExtension.appleTargets() =
